Add an test program for text viewports.
I used this program to test my fixes to VDP for the text viewport
behaviour. It should be useful to keep around.

When run, it should disable a white board and then regions filled
with a test pattern. If the whole screen is white (except for the
test pattern) then the test has failed.